Introduce SecDataDir
Starting with 2.5 mod_security requires a global
SecDataDir set for certain rules. This directory
needs to be writeable by the webuser.
It looks like EPEL Packages so far don't define such
a directory, hence we configure one for CentOS. This
requires us to also implement some settings for the
usage with itk-enabled apache. Which is very dirty at
them moment, but happens to be due to configuration
restrictions by mod_security.